+++ R/get_attribution.R
@@ -22,12 +22,12 @@
 #' @importFrom knitr combine_words
 #' @importFrom utils packageVersion
 #' @export
-#' @examples
+#' @examples \dontrun{
 #' # Get valid uuid
 #' uuid <- get_uuid(name = "Acropora cervicornis")
 #' # Get attribution data for uuid
 #' attri <- get_attribution(uuid = uuid)
-#' \dontrun{
+#' 
 #' # Get list of valid uuids
 #' uuids <- get_uuid(name = "Scleractinia", n = 5)
 #' # Get attribution data for uuids