+++ R/get_uuid.R
@@ -30,9 +30,10 @@
 #'   [pick_phylopic] to visually select the desired uuid/url.
 #' @importFrom stats setNames
 #' @export
-#' @examples
+#' @examples \dontrun{
 #' uuid <- get_uuid(name = "Acropora cervicornis")
 #' uuid <- get_uuid(name = "Dinosauria", n = 5, url = TRUE)
+#' }
 get_uuid <- function(name = NULL, img = NULL, n = 1, filter = NULL,
                      url = FALSE) {
   # Handle img -----------------------------------------------------------