+++ NEWS.md
@@ -1,3 +1,5 @@
+# rphylopic (development version)
+
 # rphylopic 1.2.2
 
 * vignettes are now precompiled